Remove Raffle
As Raffle has been replaced with the new `/random` slack command, it is
now safe to remove it from the application. This change moves the bot
user's trello tokens out of the database and into the
`TRELLO_OAUTH_TOKEN` and `TRELLO_OAUTH_SECRET` environment variables. It
also removes both the database and assets pipeline (leaving A LOT leaner
application).